% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.frame.cellResult}
\alias{as.data.frame.cellResult}
\title{Functions to coerce cellResult to data.frame}
\usage{
\method{as.data.frame}{cellResult}(x, ...)
}
\description{
Functions to coerce cellResult to data.frame
}
