% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMarker.R
\name{cellMarker}
\alias{cellMarker}
\title{assign cell type based on cell cluster results}
\usage{
cellMarker(
  x,
  type = "seurat",
  species = "human",
  keytype = "SYMBOL",
  weight = NULL,
  format = "long",
  cluster = NULL,
  tissue = NULL,
  topn = 3,
  padj = 0.05,
  minSize = 3,
  maxSize = 500,
  padj.method = "BH"
)
}
\arguments{
\item{x}{input file for marker annotation}

\item{type}{source of marker genes (seurat(default), cellranger, custom)}

\item{species}{species for annotation}

\item{keytype}{keytype for input genes}

\item{weight}{weight threshold for marker filtering}

\item{format}{file format for user supplied data.frame(long: Cluster, gene;
wide: gene, following cluster name as column name)}

\item{cluster}{clutser number (default: NULL for annotate all clusters)}

\item{tissue}{tissue for annotation (default: NULL to use all tissues)}

\item{topn}{the number of cell type to list fro each cluster}

\item{padj}{adjust p value threshold}

\item{minSize}{minimal number of genes included in significant cell type}

\item{maxSize}{maximum number of genes included in significant cell type}

\item{padj.method}{pvalue adjust method(default: "BH")}
}
\description{
assign cell type based on cell cluster results
}
\author{
Kai Guo
}
