% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cellResult-class}
\alias{cellResult-class}
\alias{show,cellResult-method}
\alias{plot,cellResult-method}
\alias{summary,cellResult-method}
\title{Class "cellResult"
This class represents the result of enrichment analysis.}
\description{
Class "cellResult"
This class represents the result of enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{enrichment analysis results}

\item{\code{detail}}{genes included in significant terms and original information}

\item{\code{species}}{species name}

\item{\code{pvalueCutoff}}{cutoff pvalue}

\item{\code{pAdjustMethod}}{pvalue adjust method}

\item{\code{padjCutoff}}{pvalue adjust cutoff value}

\item{\code{gene}}{Gene IDs}

\item{\code{keytype}}{Gene ID type}

\item{\code{sep}}{character string used to separate the genes when concatenating}
}}

\author{
Kai Guo
}
\keyword{classes}
