% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{cells}
\alias{cells}
\title{cells function for cell type identification}
\usage{
cells(
  x,
  species = "human",
  keytype = "SYMBOL",
  tissue = NULL,
  padj = 0.05,
  pvalue = NULL,
  minSize = 3,
  maxSize = 500,
  keepRich = TRUE,
  padj.method = "BH",
  sep = ","
)
}
\arguments{
\item{x}{vector contains gene names}

\item{keytype}{keytype for input genes}

\item{tissue}{tissue type (default NULL)}

\item{padj}{cutoff p adjust value}

\item{pvalue}{cutoff pvalue}

\item{minSize}{minimal number of genes included in significant terms}

\item{maxSize}{maximum number of genes included in significant terms}

\item{keepRich}{keep terms with rich factor value equal 1 or not (default: TRUE)}

\item{padj.method}{pvalue adjust method(default:"BH")}

\item{sep}{character string used to separate the genes when concatenating}

\item{sepcies}{sepcies name}
}
\description{
cells function for cell type identification
}
\author{
Kai Guo
}
