% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.getdata}
\alias{.getdata}
\title{load the data based on the species name}
\usage{
.getdata(species)
}
\arguments{
\item{species}{species name}
}
\description{
load the data based on the species name
}
\author{
Kai Guo
}
