% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getdetail}
\alias{getdetail}
\title{get detail from with cellResult and combine with other information}
\usage{
getdetail(x, y, sep)
}
\arguments{
\item{x}{cellResult object from cells function}

\item{y}{a data frame with gene name and other information}

\item{sep}{cellResult object sep}
}
\description{
get detail from with cellResult and combine with other information
}
\author{
Kai Guo
}
