% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human}
\alias{human}
\title{human}
\format{
A list of data frame with five columns individually:
\describe{
\item{SYMBOL}{SYMBOL gene IDs}
\item{cellType}{cell type}
\item{tissueType}{tissue Type}
\item{ENTREZID}{ENTREZID gene IDs}
\item{REFSEQ}{REFSEQ gene IDs}
}
}
\usage{
human
}
\description{
human cell markers for cell type identification
}
\examples{
head(human)
}
\keyword{datasets}
