% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{mouse}
\format{
A list of data frame with five columns individually:
\describe{
\item{SYMBOL}{SYMBOL gene IDs}
\item{cellType}{cell type}
\item{tissueType}{tissue Type}
\item{ENTREZID}{ENTREZID gene IDs}
\item{REFSEQ}{REFSEQ gene IDs}
}
}
\usage{
mouse
}
\description{
mouse cell markers for cell type identification
}
\examples{
head(mouse)
}
\keyword{datasets}
