% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratcells}
\alias{ratcells}
\title{rat}
\format{
A list of data frame with five columns individually:
\describe{
\item{SYMBOL}{SYMBOL gene IDs}
\item{cellType}{cell type}
\item{tissueType}{tissue Type}
\item{ENTREZID}{ENTREZID gene IDs}
\item{REFSEQ}{REFSEQ gene IDs}
\item{ENSEMBL}{ENSEMBL gene IDs}
}
}
\usage{
ratcells
}
\description{
mouse cell markers for cell type identification
}
\examples{
data(rat)
head(ratcells)
}
\keyword{datasets}
